refactor: separate gfx init in gfxApp
